more robust coding

+++ core/src/main/java/net/time4j/WeekdayInMonthElement.java
@@ -1,6 +1,6 @@
 /*
  * -----------------------------------------------------------------------
- * Copyright © 2013-2016 Meno Hochschild, <http://www.menodata.de/>
+ * Copyright © 2013-2017 Meno Hochschild, <http://www.menodata.de/>
  * -----------------------------------------------------------------------
  * This file (WeekdayInMonthElement.java) is part of project Time4J.
  *
@@ -188,7 +188,7 @@ final class WeekdayInMonthElement
 
         //~ Instanzvariablen ----------------------------------------------
 
-        private final int ordinal;
+        private final long ordinal;
         private final Weekday dayOfWeek;
         private final ChronoOperator<PlainTimestamp> specialTS;
 
@@ -233,9 +233,9 @@ final class WeekdayInMonthElement
                 Weekday current = date.get(PlainDate.DAY_OF_WEEK);
                 int delta = this.dayOfWeek.getValue() - current.getValue();
                 int dom = date.getDayOfMonth() + delta;
-                int days;
+                long days;
                 if (this.ordinal == LAST) {
-                    days = (5 - (Math.floorDiv(dom - 1, 7) + 1)) * 7 + delta;
+                    days = (5L - (Math.floorDiv(dom - 1, 7) + 1)) * 7 + delta;
                     int max = GregorianMath.getLengthOfMonth(date.getYear(), date.getMonth());
                     if (date.getDayOfMonth() + days > max) {
                         days -= 7;
